requirements.txt
opencv-python-headless
numpy
Pillow
python-docx
pdf2image